Move integration tests to the same module
