--- tests/integration/test_cos_integration.py
@@ -1,93 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2023 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""Integration tests for Synapse charm COS integration."""
-import logging
-import typing
-
-import pytest
-import requests
-from juju.model import Model
-from ops.model import ActiveStatus
-
-# caused by pytest fixtures
-# pylint: disable=too-many-arguments
-
-# mypy has trouble to inferred types for variables that are initialized in subclasses.
-ACTIVE_STATUS_NAME = typing.cast(str, ActiveStatus.name)  # type: ignore
-
-logger = logging.getLogger(__name__)
-
-
-@pytest.mark.usefixtures("synapse_app")
-async def test_prometheus_integration(
-    model: Model,
-    prometheus_app_name: str,
-    synapse_app_name: str,
-    prometheus_app,  # pylint: disable=unused-argument
-    get_unit_ips: typing.Callable[[str], typing.Awaitable[tuple[str, ...]]],
-):
-    """
-    arrange: after Synapse charm has been deployed.
-    act: establish relations with prometheus charm.
-    assert: prometheus metrics endpoint for prometheus is active and prometheus has active scrape
-        targets.
-    """
-    await model.add_relation(prometheus_app_name, synapse_app_name)
-    await model.wait_for_idle(
-        apps=[synapse_app_name, prometheus_app_name], status=ACTIVE_STATUS_NAME
-    )
-
-    for unit_ip in await get_unit_ips(prometheus_app_name):
-        query_targets = requests.get(f"http://{unit_ip}:9090/api/v1/targets", timeout=10).json()
-        assert len(query_targets["data"]["activeTargets"])
-
-
-@pytest.mark.usefixtures("synapse_app")
-async def test_grafana_integration(
-    model: Model,
-    synapse_app_name: str,
-    prometheus_app_name: str,
-    prometheus_app,  # pylint: disable=unused-argument
-    grafana_app_name: str,
-    grafana_app,  # pylint: disable=unused-argument
-    get_unit_ips: typing.Callable[[str], typing.Awaitable[tuple[str, ...]]],
-):
-    """
-    arrange: after Synapse charm has been deployed.
-    act: establish relations with grafana charm.
-    assert: grafana Synapse dashboard can be found.
-    """
-    await model.relate(
-        f"{prometheus_app_name}:grafana-source", f"{grafana_app_name}:grafana-source"
-    )
-    await model.relate(synapse_app_name, grafana_app_name)
-
-    await model.wait_for_idle(
-        apps=[synapse_app_name, prometheus_app_name, grafana_app_name],
-        status=ACTIVE_STATUS_NAME,
-        idle_period=60,
-    )
-
-    action = await model.applications[grafana_app_name].units[0].run_action("get-admin-password")
-    await action.wait()
-    password = action.results["admin-password"]
-    grafana_ip = (await get_unit_ips(grafana_app_name))[0]
-    sess = requests.session()
-    sess.post(
-        f"http://{grafana_ip}:3000/login",
-        json={
-            "user": "admin",
-            "password": password,
-        },
-    ).raise_for_status()
-    datasources = sess.get(f"http://{grafana_ip}:3000/api/datasources", timeout=10).json()
-    datasource_types = set(datasource["type"] for datasource in datasources)
-    assert "prometheus" in datasource_types
-    dashboards = sess.get(
-        f"http://{grafana_ip}:3000/api/search",
-        timeout=10,
-        params={"query": "Synapse Operator"},
-    ).json()
-    assert len(dashboards)
